requirements.txt
iotlabcli>=2.5
PyYAML>=5.4
